+++ amazonorders/orders.py
@@ -73,8 +73,8 @@ class AmazonOrders:
             response_parsed = self.amazon_session.last_response_parsed
 
             for order_tag in util.select(response_parsed, self.config.selectors.ORDER_HISTORY_ENTITY_SELECTOR):
-                # First check if this Order is known to be of a type that we do not currently have a way to support parsing,
-                # meaning it should be skipped
+                # First check if this Order is known to be of a type that we do not currently have a way to support
+                # parsing, meaning it should be skipped
                 if util.select(order_tag, self.config.selectors.ORDER_HISTORY_BRAND_SELECTOR):
                     continue